requirements.txt
Flask==3.0.3
gunicorn==23.0.0
eventlet==0.36.1
twilio==9.3.6
python-dotenv==1.0.1
websockets==12.0
simple-websocket==1.0.0
requests==2.32.3
